+++ src/pages/Product/BycleInfo/BycleInfo.scss
@@ -7,6 +7,7 @@
   border-radius: 1rem;
   background-color: white;
   cursor: pointer;
+
   &:hover {
     box-shadow: 5px 5px 5px 5px gray;
     transition-duration: 0.8s;
@@ -15,27 +16,33 @@
       transition-duration: 0.8s;
     }
   }
+
   .bycleContent {
     display: flex;
     flex-direction: column;
     align-items: center;
     margin: 1.5rem 0;
+
     .bycleName {
       font-size: 3rem;
       margin-bottom: 1.5rem;
     }
+
     .byclePrice {
       font-size: 2rem;
       margin-bottom: 1.5rem;
     }
+
     .bycleHashtag {
       display: flex;
       flex-wrap: wrap;
       justify-content: center;
-      button {
+
+      .hashTag {
         background-color: rgb(231, 231, 231);
         border: none;
         margin: 0.7rem 0.7rem;
+
         &:hover {
           box-shadow: 0.5rem 0.5rem 0.5rem 0.5rem gray;
           transition-duration: 0.8s;
